requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
Pillow
dj-database-url
whitenoise
django-environ
gunicorn
openpyxl